Move common code to AbstractHttpServiceAgentSession
